server/requirements.txt
Flask-SQLAlchemy
passlib
psycopg2-binary

Flask-RESTful
Flask-jwt-extended

flask-cors